Move Views and Components from Folder Compose to Screens
