+++ budget-binder-multiplatform-app/src/jsMain/kotlin/de/hsfl/budgetBinder/compose/Router.kt
@@ -3,13 +3,13 @@ package de.hsfl.budgetBinder.compose
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.rememberCoroutineScope
-import de.hsfl.budgetBinder.compose.dashboard.DashboardComponent
-import de.hsfl.budgetBinder.compose.entry.EntryComponent
-import de.hsfl.budgetBinder.compose.login.LoginComponent
-import de.hsfl.budgetBinder.compose.register.RegisterComponent
-import de.hsfl.budgetBinder.compose.settings.SettingsComponent
 import de.hsfl.budgetBinder.presentation.Screen
 import de.hsfl.budgetBinder.presentation.flow.RouterFlow
+import de.hsfl.budgetBinder.screens.dashboard.DashboardComponent
+import de.hsfl.budgetBinder.screens.entry.EntryComponent
+import de.hsfl.budgetBinder.screens.login.LoginComponent
+import de.hsfl.budgetBinder.screens.register.RegisterComponent
+import de.hsfl.budgetBinder.screens.settings.SettingsComponent
 import di
 import org.jetbrains.compose.web.dom.Text
 import org.kodein.di.instance
