+++ budget-binder-multiplatform-app/src/jsMain/kotlin/de/hsfl/budgetBinder/screens/category/CategoryCreateOnRegisterView.kt
@@ -1,4 +1,4 @@
-package de.hsfl.budgetBinder.compose.category
+package de.hsfl.budgetBinder.screens.category
 
 import androidx.compose.runtime.*
 import de.hsfl.budgetBinder.presentation.UiState
