+++ budget-binder-multiplatform-app/src/jsMain/kotlin/de/hsfl/budgetBinder/screens/category/CategoryEditView.kt
@@ -1,4 +1,4 @@
-package de.hsfl.budgetBinder.compose.category
+package de.hsfl.budgetBinder.screens.category
 
 import androidx.compose.runtime.*
 import de.hsfl.budgetBinder.common.Category
@@ -15,13 +15,14 @@ import org.jetbrains.compose.web.dom.*
 
 
 @Composable
-fun CategoryCreateView(
+fun CategoryEditView(
     state: State<Any>,
-    onCreateCategoryButtonPressed: (name: String, color: String, image: Category.Image, budget: Float) -> Unit,
+    onEditCategoryButtonPressed: (name: String, color: String, image: Category.Image, budget: Float) -> Unit,
     onChangeToDashboard: () -> Unit,
-    onChangeToSettings: () -> Unit,
     onChangeToCategory: () -> Unit,
+    onChangeToSettings: () -> Unit
 ) {
+    var category by remember { mutableStateOf(Category(0, "", "", Category.Image.DEFAULT, 0f)) }
     var categoryNameTextFieldState by remember { mutableStateOf("") }
     var categoryColorTextFieldState by remember { mutableStateOf("") }
     var categoryImageState = remember { mutableStateOf(Category.Image.DEFAULT) }
@@ -68,15 +69,12 @@ fun CategoryCreateView(
         })
 
     MainFlexContainer {
-        H1(
-            attrs = {
-                style { margin(2.percent) }
-            }
-        ) { Text("Create a new Category") }
+        H1(attrs = { style { margin(2.percent) } }) { Text("Edit Category") }
+
         Form(attrs = {
             this.addEventListener("submit") {
                 console.log("$categoryNameTextFieldState, $categoryColorTextFieldState, $categoryImageState, $categoryBudgetTextFieldState")
-                onCreateCategoryButtonPressed(
+                onEditCategoryButtonPressed(
                     categoryNameTextFieldState,
                     categoryColorTextFieldState,
                     categoryImageState.value,
@@ -236,10 +234,19 @@ fun CategoryCreateView(
                         value("Submit")
                     })
             }
+
             Div {
                 when (viewState) {
                     is UiState.Success<*> -> {
-                        //Text((viewState as UiState.Success<*>).element.toString())
+                        when (val element = (viewState as UiState.Success<*>).element) {
+                            is Category -> {
+                                category = element
+                                categoryNameTextFieldState = category.name
+                                categoryColorTextFieldState = "#" + category.color
+                                categoryImageState.value = category.image
+                                categoryBudgetTextFieldState = category.budget.toString()
+                            }
+                        }
                     }
                     is UiState.Error -> {
                         Text((viewState as UiState.Error).error)
