+++ budget-binder-multiplatform-app/src/jsMain/kotlin/de/hsfl/budgetBinder/screens/dashboard/DashboardComponent.kt
@@ -1,16 +1,15 @@
-package de.hsfl.budgetBinder.compose.dashboard
+package de.hsfl.budgetBinder.screens.dashboard
 
 import androidx.compose.runtime.*
 import de.hsfl.budgetBinder.common.Category
+import de.hsfl.budgetBinder.compose.BudgetBar
 import de.hsfl.budgetBinder.compose.Icon
 import de.hsfl.budgetBinder.compose.MainFlexContainer
 import de.hsfl.budgetBinder.compose.NavBar
-import de.hsfl.budgetBinder.compose.category.BudgetBar
 import de.hsfl.budgetBinder.compose.theme.AppStylesheet
 import de.hsfl.budgetBinder.presentation.CategoryImageToIcon
 import de.hsfl.budgetBinder.presentation.event.LifecycleEvent
 import de.hsfl.budgetBinder.presentation.event.UiEvent
-import de.hsfl.budgetBinder.presentation.viewmodel.auth.register.RegisterEvent
 import de.hsfl.budgetBinder.presentation.viewmodel.dashboard.DashboardEntryState
 import de.hsfl.budgetBinder.presentation.viewmodel.dashboard.DashboardEvent
 import de.hsfl.budgetBinder.presentation.viewmodel.dashboard.DashboardState
