+++ budget-binder-multiplatform-app/src/jsMain/kotlin/de/hsfl/budgetBinder/screens/entry/EntryEditView.kt
@@ -1,13 +1,16 @@
-package de.hsfl.budgetBinder.compose.entry
+package de.hsfl.budgetBinder.screens.entry
 
 import androidx.compose.runtime.*
-import de.hsfl.budgetBinder.compose.*
+import de.hsfl.budgetBinder.compose.ChooseCategoryMenu
 import de.hsfl.budgetBinder.compose.theme.AppStylesheet
 import de.hsfl.budgetBinder.presentation.viewmodel.entry.EntryEvent
 import de.hsfl.budgetBinder.presentation.viewmodel.entry.EntryViewModel
 import di
 import org.jetbrains.compose.web.ExperimentalComposeWebSvgApi
-import org.jetbrains.compose.web.attributes.*
+import org.jetbrains.compose.web.attributes.ButtonType
+import org.jetbrains.compose.web.attributes.InputType
+import org.jetbrains.compose.web.attributes.required
+import org.jetbrains.compose.web.attributes.type
 import org.jetbrains.compose.web.css.*
 import org.jetbrains.compose.web.dom.*
 import org.jetbrains.compose.web.svg.Path
@@ -17,26 +20,29 @@ import org.kodein.di.instance
 
 @OptIn(ExperimentalComposeWebSvgApi::class)
 @Composable
-fun EntryCreateView(
-    onCreateButton: () -> Unit,
+fun EntryEditView(
+    onEditButton: () -> Unit,
 ) {
     val viewModel: EntryViewModel by di.instance()
     //Input
     val entryNameTextField by viewModel.nameText.collectAsState()
     val entryAmountTextField by viewModel.amountText.collectAsState()
+    val entryRepeat by viewModel.repeatState.collectAsState()
     val entryCategoryIDTextField by viewModel.categoryIDState.collectAsState()
     val amountSign by viewModel.amountSignState.collectAsState()
     //Data
     val categoryList by viewModel.categoryListState.collectAsState()
-
+    val entry by viewModel.selectedEntryState.collectAsState()
+    console.log("unser Entry ist $entry")
+    console.log("$entryNameTextField und $entryAmountTextField und $entryCategoryIDTextField")
     H1(
         attrs = {
             style { margin(2.percent) }
         }
-    ) { Text("Create new Entry") }
+    ) { Text("Edit Entry") }
     Form(attrs = {
         this.addEventListener("submit") {
-            onCreateButton()
+            onEditButton()
             it.preventDefault()
         }
     }
@@ -147,10 +153,10 @@ fun EntryCreateView(
                     type = InputType.Number
                 ) {
                     attr("step", "0.01")
-                    value(entryAmountTextField)
                     classes("mdc-text-field__input")
+                    value(entryAmountTextField)
                     onInput {
-                        viewModel.onEvent(EntryEvent.EnteredAmount(it.value!!.toFloat()))
+                        viewModel.onEvent(EntryEvent.EnteredAmount(it.value as Float))
                     }
                 }
                 Span(
@@ -168,14 +174,15 @@ fun EntryCreateView(
             Div(attrs = { style { flex(50.percent) } }) {
                 Div(attrs = { classes("mdc-form-field") }) {
                     Div(attrs = { classes("mdc-checkbox") }) {
-                        Input(type = InputType.Checkbox)
+                        CheckboxInput (attrs =
                         {
+                            checked(entryRepeat)
                             classes("mdc-checkbox__native-control")
                             id("checkbox-1")
                             onInput {
                                 viewModel.onEvent(EntryEvent.EnteredRepeat)
                             }
-                        }
+                        })
                         Div(attrs = { classes("mdc-checkbox__background") }) {
                             Svg(
                                 attrs = { classes("mdc-checkbox__checkmark") },
@@ -192,7 +199,10 @@ fun EntryCreateView(
                     Label(forId = "checkbox-1") { Text("repeat") }
                 }
             }
-            Div(attrs = { style { flex(50.percent) } }) {
+            Div(attrs = { style {
+                flex(50.percent)
+                alignItems(AlignItems.Stretch)
+            } }) {
                 ChooseCategoryMenu(categoryList, entryCategoryIDTextField) { id ->
                     viewModel.onEvent(EntryEvent.EnteredCategoryID(id))
                 }
@@ -203,16 +213,6 @@ fun EntryCreateView(
                 classes(AppStylesheet.margin)
             }
         ) {
-            Button(
-                attrs = {
-                    classes("mdc-button", "mdc-button--raised")
-                    type(ButtonType.Button)
-                    onClick { viewModel.onEvent(EntryEvent.OnCancel) }
-                }
-            ) {
-                Span(attrs = { classes("mdc-button__label") }
-                ) { Text("Cancel") }
-            }
             SubmitInput(
                 attrs = {
                     classes("mdc-button", "mdc-button--raised")
@@ -221,4 +221,3 @@ fun EntryCreateView(
         }
     }
 }
-
