+++ budget-binder-multiplatform-app/src/jsMain/kotlin/de/hsfl/budgetBinder/screens/entry/EntryOverviewView.kt
@@ -1,4 +1,4 @@
-package de.hsfl.budgetBinder.compose.entry
+package de.hsfl.budgetBinder.screens.entry
 
 import androidx.compose.runtime.*
 import de.hsfl.budgetBinder.common.Entry
