+++ budget-binder-multiplatform-app/src/jsMain/kotlin/de/hsfl/budgetBinder/screens/login/LoginComponent.kt
@@ -1,4 +1,4 @@
-package de.hsfl.budgetBinder.compose.login
+package de.hsfl.budgetBinder.screens.login
 
 import androidx.compose.runtime.*
 import de.hsfl.budgetBinder.compose.MainFlexContainer
